+++ changelog.md
@@ -3,5 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## 0.1.0 [2018-09-30]
+## Unreleased 0.2.0 [2018-09-30]
+- Database Generation
+- Console Logs Look Nicer
+
+## Unreleased 0.1.0 [2018-09-30]
 - Start Changelog